+++ core/stencil.config.ts
@@ -87,23 +87,18 @@ export const config: Config = {
         'ion-app',
         'ion-icon'
       ],
-      routerLinkComponents: [
-        'ion-card',
-        'ion-item',
-        'ion-button',
-        'ion-fab-button',
-
-      ],
       componentModels: [
         {
           elements: ['ion-checkbox', 'ion-toggle'],
           targetAttr: 'checked',
-          event: 'ionChange'
+          event: 'v-ionChange',
+          externalEvent: 'ionChange'
         },
         {
           elements: ['ion-datetime', 'ion-input', 'ion-radio-group', 'ion-radio', 'ion-range', 'ion-searchbar', 'ion-segment', 'ion-segment-button', 'ion-select', 'ion-textarea'],
           targetAttr: 'value',
-          event: 'ionChange'
+          event: 'v-ionChange',
+          externalEvent: 'ionChange'
         }
       ],
     }),