+++ packages/react-router/test-app/src/pages/routing/Routing.tsx
@@ -1,5 +1,13 @@
 import React from 'react';
-import { IonContent, IonHeader, IonPage, IonTitle, IonToolbar, IonRouterOutlet, IonSplitPane } from '@ionic/react';
+import {
+  IonContent,
+  IonHeader,
+  IonPage,
+  IonTitle,
+  IonToolbar,
+  IonRouterOutlet,
+  IonSplitPane,
+} from '@ionic/react';
 import Menu from './Menu';
 import { Route, Redirect } from 'react-router';
 import Tabs from './Tabs';
@@ -8,8 +16,7 @@ import OtherPage from './OtherPage';
 import PropsTest from './PropsTest';
 import RedirectRouting from './RedirectRouting';
 
-interface RoutingProps {
-}
+interface RoutingProps {}
 
 const Routing: React.FC<RoutingProps> = () => {
   return (
@@ -38,9 +45,16 @@ const Routing: React.FC<RoutingProps> = () => {
         <Route path="/routing/propstest" component={PropsTest} />
         <Route path="/routing/redirect" render={() => <Redirect to="/routing/tabs" />} />
         <Route path="/routing/redirect-routing" render={() => <RedirectRouting />} />
-        <Route render={() => <IonPage data-pageid="not-found"><IonContent><div>Not found</div></IonContent></IonPage>} />
+        <Route
+          render={() => (
+            <IonPage data-pageid="not-found">
+              <IonContent>
+                <div>Not found</div>
+              </IonContent>
+            </IonPage>
+          )}
+        />
         {/* <Route render={() => <Redirect to="/tabs" />} /> */}
-
       </IonRouterOutlet>
     </IonSplitPane>
   );