+++ packages/react/src/components/IonRoute.tsx
@@ -10,30 +10,25 @@ export interface IonRouteProps {
   disableIonPageManagement?: boolean;
 }
 
-interface IonRouteState {
-
-}
+interface IonRouteState {}
 
 export class IonRoute extends React.PureComponent<IonRouteProps, IonRouteState> {
-
   context!: React.ContextType<typeof NavContext>;
 
   render() {
-
     const IonRouteInner = this.context.getIonRoute();
 
     if (!this.context.hasIonicRouter() || !IonRoute) {
-      console.error('You either do not have an Ionic Router package, or your router does not support using <IonRoute>');
+      console.error(
+        'You either do not have an Ionic Router package, or your router does not support using <IonRoute>'
+      );
       return null;
     }
 
-    return (
-      <IonRouteInner {...this.props} />
-    );
+    return <IonRouteInner {...this.props} />;
   }
 
   static get contextType() {
     return NavContext;
   }
-
 }