+++ packages/vue/src/components/IonRouterOutlet.ts
@@ -11,19 +11,27 @@ import {
   onUnmounted
 } from 'vue';
 import { AnimationBuilder, LIFECYCLE_DID_ENTER, LIFECYCLE_DID_LEAVE, LIFECYCLE_WILL_ENTER, LIFECYCLE_WILL_LEAVE } from '@ionic/core';
-import { matchedRouteKey, useRoute } from 'vue-router';
+import { matchedRouteKey, routeLocationKey, useRoute } from 'vue-router';
 import { fireLifecycle, generateId, getConfig } from '../utils';
 
 let viewDepthKey: InjectionKey<0> = Symbol(0);
 export const IonRouterOutlet = defineComponent({
   name: 'IonRouterOutlet',
   setup(_, { attrs }) {
+    const injectedRoute = inject(routeLocationKey)!;
     const route = useRoute();
     const depth = inject(viewDepthKey, 0);
+    let usingDeprecatedRouteSetup = false;
+
+    // TODO: Remove in Ionic Vue v6.0
+    if (attrs.tabs && route.matched[depth]?.children?.length > 0) {
+      console.warn('[@ionic/vue Deprecation]: Your child routes are nested inside of each tab in your routing config. This format will not be supported in Ionic Vue v6.0. Instead, write your child routes as sibling routes. See https://ionicframework.com/docs/vue/navigation#child-routes-within-tabs for more information.');
+      usingDeprecatedRouteSetup = true;
+    }
     const matchedRouteRef: any = computed(() => {
       const matchedRoute = route.matched[depth];
 
-      if (matchedRoute && attrs.tabs && route.matched[depth + 1]) {
+      if (matchedRoute && attrs.tabs && route.matched[depth + 1] && usingDeprecatedRouteSetup) {
         return route.matched[route.matched.length - 1];
       }
 
@@ -75,15 +83,15 @@ export const IonRouterOutlet = defineComponent({
        * to make sure the view is in the outlet we want.
        */
       const routeInfo = ionRouter.getCurrentRouteInfo();
-      const enteringViewItem = viewStacks.findViewItemByRouteInfo({ pathname: routeInfo.pushedByRoute }, id);
+      const enteringViewItem = viewStacks.findViewItemByRouteInfo({ pathname: routeInfo.pushedByRoute || '' }, id, usingDeprecatedRouteSetup);
 
       return !!enteringViewItem;
     }
     const onStart = async () => {
       const routeInfo = ionRouter.getCurrentRouteInfo();
       const { routerAnimation } = routeInfo;
-      const enteringViewItem = viewStacks.findViewItemByRouteInfo({ pathname: routeInfo.pushedByRoute }, id);
-      const leavingViewItem = viewStacks.findViewItemByRouteInfo(routeInfo, id);
+      const enteringViewItem = viewStacks.findViewItemByRouteInfo({ pathname: routeInfo.pushedByRoute || '' }, id, usingDeprecatedRouteSetup);
+      const leavingViewItem = viewStacks.findViewItemByRouteInfo(routeInfo, id, usingDeprecatedRouteSetup);
 
       if (leavingViewItem) {
         let animationBuilder = routerAnimation;
@@ -138,7 +146,7 @@ export const IonRouterOutlet = defineComponent({
          * re-hide the page that was going to enter.
          */
         const routeInfo = ionRouter.getCurrentRouteInfo();
-        const enteringViewItem = viewStacks.findViewItemByRouteInfo({ pathname: routeInfo.pushedByRoute }, id);
+        const enteringViewItem = viewStacks.findViewItemByRouteInfo({ pathname: routeInfo.pushedByRoute || '' }, id, usingDeprecatedRouteSetup);
         enteringViewItem.ionPageElement.setAttribute('aria-hidden', 'true');
         enteringViewItem.ionPageElement.classList.add('ion-page-hidden');
       }
@@ -191,17 +199,21 @@ export const IonRouterOutlet = defineComponent({
 
     const handlePageTransition = async () => {
       const routeInfo = ionRouter.getCurrentRouteInfo();
-      const { routerDirection, routerAction, routerAnimation } = routeInfo;
+      const { routerDirection, routerAction, routerAnimation, prevRouteLastPathname } = routeInfo;
 
-      const enteringViewItem = viewStacks.findViewItemByRouteInfo(routeInfo, id);
-      const leavingViewItem = viewStacks.findLeavingViewItemByRouteInfo(routeInfo, id);
+      const enteringViewItem = viewStacks.findViewItemByRouteInfo(routeInfo, id, usingDeprecatedRouteSetup);
+      let leavingViewItem = viewStacks.findLeavingViewItemByRouteInfo(routeInfo, id, true, usingDeprecatedRouteSetup);
       const enteringEl = enteringViewItem.ionPageElement;
 
       if (enteringViewItem === leavingViewItem) return;
 
+      if (!leavingViewItem && prevRouteLastPathname) {
+        leavingViewItem = viewStacks.findViewItemByPathname(prevRouteLastPathname, id, usingDeprecatedRouteSetup);
+      }
+
       fireLifecycle(enteringViewItem.vueComponent, enteringViewItem.vueComponentRef, LIFECYCLE_WILL_ENTER);
 
-      if (leavingViewItem) {
+      if (leavingViewItem && enteringViewItem !== leavingViewItem) {
         let animationBuilder = routerAnimation;
         const leavingEl = leavingViewItem.ionPageElement;
 
@@ -291,7 +303,7 @@ export const IonRouterOutlet = defineComponent({
       }
 
       const currentRoute = ionRouter.getCurrentRouteInfo();
-      let enteringViewItem = viewStacks.findViewItemByRouteInfo(currentRoute, id);
+      let enteringViewItem = viewStacks.findViewItemByRouteInfo(currentRoute, id, usingDeprecatedRouteSetup);
 
       if (!enteringViewItem) {
         enteringViewItem = viewStacks.createViewItem(id, matchedRouteRef.value.components.default, matchedRouteRef.value, currentRoute);
@@ -352,12 +364,13 @@ export const IonRouterOutlet = defineComponent({
     return {
       id,
       components,
+      injectedRoute,
       ionRouterOutlet,
       registerIonPage
     }
   },
   render() {
-    const { components, registerIonPage } = this;
+    const { components, registerIonPage, injectedRoute } = this;
 
     return h(
       'ion-router-outlet',
@@ -374,22 +387,43 @@ export const IonRouterOutlet = defineComponent({
         /**
          * IonRouterOutlet does not support named outlets.
          */
-        if (c.matchedRoute?.props?.default) {
-          const matchedRoute = c.matchedRoute;
-          const routePropsOption = matchedRoute.props.default;
-          const routeProps = routePropsOption
-            ? routePropsOption === true
-              ? c.params
-              : typeof routePropsOption === 'function'
-              ? routePropsOption(matchedRoute)
-              : routePropsOption
-            : null
+        const routePropsOption = c.matchedRoute?.props?.default;
 
-          props = {
-            ...props,
-            ...routeProps
+        /**
+         * Since IonRouterOutlet renders multiple components,
+         * each render will cause all props functions to be
+         * called again. As a result, we need to cache the function
+         * result and provide it on each render so that the props
+         * are not lost when navigating from and back to a page.
+         * When a component is destroyed and re-created, the
+         * function is called again.
+         */
+        const getPropsFunctionResult = () => {
+          const cachedPropsResult = c.vueComponentData?.propsFunctionResult;
+          if (cachedPropsResult) {
+            return cachedPropsResult;
+          } else {
+            const propsFunctionResult = routePropsOption(injectedRoute);
+            c.vueComponentData = {
+              ...c.vueComponentData,
+              propsFunctionResult
+            };
+            return propsFunctionResult;
           }
         }
+        const routeProps = routePropsOption
+          ? routePropsOption === true
+            ? c.params
+            : typeof routePropsOption === 'function'
+            ? getPropsFunctionResult()
+            : routePropsOption
+          : null
+
+        props = {
+          ...props,
+          ...routeProps
+        }
+
         return h(
           c.vueComponent,
           props